[Android] Add support for removeJavascriptInterface
Native side already has the implementation, so it only
need to be exposed in XWalkView.
Test case has also be changed to adapt this new API.

BUG=XWALK-6882
