+++ test/android/data/add_js_interface.html
@@ -1,5 +1,8 @@
 <html>
 <head>
+<title>
+Add JS Interface
+</title>
 <script type="text/javascript">
     function setTitle()
     {
